Extract `StableAssociativeCollection` out of `StableMap`
