+++ DevToolkt/core/src/commonMain/kotlin/dev/toolkt/core/collections/MutableTreeMap.kt
@@ -1,6 +1,6 @@
 package dev.toolkt.core.collections
 
-import dev.toolkt.core.collections.StableMap.Handle
+import dev.toolkt.core.collections.StableAssociativeCollection.Handle
 import dev.toolkt.core.data_structures.binary_tree.BinaryTree
 import dev.toolkt.core.data_structures.binary_tree.RedBlackTree
 import dev.toolkt.core.data_structures.binary_tree.findBy
